macro cleanup
