+++ common/src/parsing.rs
@@ -13,29 +13,11 @@
 // limitations under the License.
 
 use anyhow::{Error, Result};
+use aoc_solution::parser::AocInputParser;
 use std::fmt::Debug;
 use std::marker::PhantomData;
 use std::ops::RangeInclusive;
 use std::str::FromStr;
-
-// we need separate trait, i.e. we can't just use `FromStr`,
-// because of all the custom rules for say `Vec<T>`
-// TODO: or maybe we should just create wrapper containers instead?
-pub trait AocInputParser {
-    type Output;
-
-    fn parse_input(raw: &str) -> Result<Self::Output>;
-}
-
-pub trait AocParseExt {
-    fn parse_aoc_input<F: AocInputParser>(&self) -> Result<F::Output>;
-}
-
-impl AocParseExt for str {
-    fn parse_aoc_input<F: AocInputParser>(&self) -> Result<F::Output> {
-        <F as AocInputParser>::parse_input(self)
-    }
-}
 
 pub struct GroupsParser<T>(*const PhantomData<T>);
